backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
pymongo==4.13.0
python-dotenv==1.1.0
bcrypt==4.3.0
PyJWT==2.10.1
python-multipart==0.0.20
email-validator==2.2.0
python-jose[cryptography]==3.5.0
